--- paddle/cinn/backends/codegen_host_base.h
@@ -1,57 +1,0 @@
-// Copyright (c) 2024 PaddlePaddle Authors. All Rights Reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#pragma once
-
-#include <memory>
-
-#include "paddle/cinn/backends/llvm/codegen_llvm.h"
-#include "paddle/cinn/runtime/intrinsic.h"
-
-namespace cinn {
-namespace backends {
-
-/**
- * CodeGenHostBase takes a CINN Module with host functions and output a LLVM
- * module.
- */
-class CodeGenHostBase : public CodeGenLLVM {
- public:
-  explicit CodeGenHostBase(llvm::Module *m,
-                           llvm::IRBuilder<> *b,
-                           const std::shared_ptr<SymbolTable> &vars = nullptr)
-      : CodeGenLLVM(m, b, vars) {}
-
-  using CodeGenLLVM::Visit;
-  llvm::Value *Visit(const ir::_LoweredFunc_ *func) override {
-    return LowerHostFunc(func);
-  }
-
-  llvm::Value *Visit(const ir::Call *op) override {
-    // TODO(Hongqing-work): change intrinsic name to get_value_in_kernel_args
-    if (op->name == runtime::intrinsic::get_value_in_cuda_kernel_args) {
-      return LowerParseArgsValueCall(op);
-    } else {
-      return CodeGenLLVM::Visit(op);
-    }
-  }
-
- protected:
-  llvm::Value *LowerHostFunc(const ir::_LoweredFunc_ *func);
-
-  llvm::Value *LowerParseArgsValueCall(const ir::Call *call_ir);
-};
-
-}  // namespace backends
-}  // namespace cinn
